+++ packages/backend/src/services/DashboardService/DashboardService.mock.ts
@@ -87,6 +87,7 @@ export const dashboard: Dashboard = {
     },
     spaceUuid: 'spaceUuid',
     spaceName: 'space name',
+    pinnedListUuid: undefined,
 };
 
 export const dashboardsDetails: DashboardBasicDetails[] = [
@@ -98,6 +99,7 @@ export const dashboardsDetails: DashboardBasicDetails[] = [
         description: dashboard.description,
         updatedAt: dashboard.updatedAt,
         spaceUuid: 'spaceUuid',
+        pinnedListUuid: undefined,
     },
 ];