+++ packages/common/src/types/savedCharts.ts
@@ -18,6 +18,7 @@ export enum ChartType {
     TABLE = 'table',
     BIG_NUMBER = 'big_number',
 }
+
 export type BigNumber = {
     label?: string;
     style?: CompactOrAlias;
@@ -210,6 +211,7 @@ export type SavedChart = {
     spaceUuid: string;
     spaceName: string;
     views?: number;
+    pinnedListUuid: string | undefined;
 };
 
 export type CreateSavedChart = Omit<
@@ -220,6 +222,7 @@ export type CreateSavedChart = Omit<
     | 'organizationUuid'
     | 'spaceUuid'
     | 'spaceName'
+    | 'pinnedListUuid'
 > & { spaceUuid?: string };
 
 export type CreateSavedChartVersion = Omit<
@@ -231,11 +234,11 @@ export type CreateSavedChartVersion = Omit<
     | 'organizationUuid'
     | 'spaceUuid'
     | 'spaceName'
+    | 'pinnedListUuid'
 >;
 
-export type UpdateSavedChart = Pick<
-    SavedChart,
-    'name' | 'description' | 'spaceUuid'
+export type UpdateSavedChart = Partial<
+    Pick<SavedChart, 'name' | 'description' | 'spaceUuid'>
 >;
 
 export type UpdateMultipleSavedChart = Pick<
@@ -252,6 +255,7 @@ export type SpaceQuery = Pick<
     | 'description'
     | 'spaceUuid'
     | 'views'
+    | 'pinnedListUuid'
 > & { chartType?: ChartKind | undefined };
 
 export const isCompleteLayout = (